Addressing issue #102.
Ansible 2 adds a feature to search the modules return value for any
strings which should not be logged. This feature only handles built-in
data types and does not handle custom objects.

The junos_get_facts module was returning a junos.version_info object
which caused this new feature to raise a TypeError.

Since all Ansible return values are passed via JSON, the junos.version_info
object was previously getting converted to a dictionary anyway, so this
change should be backwards compatible with Ansible 1 as well.