JSON dump of Junos facts requires native Python data types. Fixes #223 (#224)
After Juniper/py-junos-eznc#638, PyEZ facts are now a custom object
rather than a true dict. In addition, facts are now read-only.

This was partially addressed in #210, but I missed the case
where the `savedir` option is specified to dump the facts into
JSON files. The same fix is also required in this case.