Add the force_host argument to junos_install_os. Addresses #233. (#234)
* Preparing for v.1.4.1 release.

* Add the `force_host` argument to `junos_install_os`. Addresses #233.

By default, the Host Software package on QFX-series platforms is not
upgraded by the <request-package-add> RPC. Adding the <force-host/>
argument to the RPC forces the Host Software package to also be upgraded.

This PR adds the ability to pass the `force_host` argument to the
`junos_install_os` module. This optional boolean argument defaults
to False. When set to True, it passes the <force-host/> argument to
the underlying <request-package-add> RPC.

* Removing Ansible 1.9.6 from testing.