Mimic old behavior for junos_install_config with check_commit = False.
There was a backwards compatibility issue with junos_install_config.

junos_install_config:
   check_commit: false

Before the above used to perform a diff, but not perform a "commit check"
or a "commit". Now, it performs the diff, but it also does a "commit".

This change restores the previous behavior where:

junos_install_config:
   check_commit: false

will only perform the diff.