Handle empty 'version_info' and 'junos_info' facts more gracefully. Addresses #360.
When PyEZ is unable to gather a fact from a Junos device, the key still exists
in the dev.facts attribute, but has a value of None. The checks for converting
the `version_info` and `junos_info` facts from their custom types to a dict
did not properly account for this situation. IN that case, they would attempt
to convert `None` to a `dict` and throw an exception.

This PR more gracefully handles that situation. However, it does not fix the
underlying Junos or PyEZ issue which might lead to being unable to collect
the `version_info` and `junos_info` facts from a device. That would need
to be addresses separately in PyEZ.