Fixed rpd module related issue
When performing an RPC get with an RPC that is not implemented for a platform, it fails when using a PyEZ connection.
With a local connection, the response is “Unable to execute the RPC” and it continues with the next RPC get.
PyEZ behavior should be the same as with the local connection.